ISSUE #2258: Moved form javascript components to new sub directory
