+++ public/js/admin.js
@@ -1,11 +1,11 @@
-import initPasswordToggle from "./components/password-toggle";
-import initFormLoadingState from "./components/form-loading-state";
-import initDispatchCommandForm from "./components/dispatch-command-form";
-import initDependentSelects from "./components/dependent-select";
-import initDependentFormInputs from "./components/dependent-form-input";
+import initPasswordToggle from "./components/form/password-toggle";
+import initFormLoadingState from "./components/form/form-loading-state";
+import initDispatchCommandForm from "./components/form/dispatch-command-form";
+import initDependentSelects from "./components/form/dependent-select";
+import initDependentFormInputs from "./components/form/dependent-form-input";
 import initDropdowns from "./components/dropdown";
 import initRebuildStatus from "./components/rebuild-status";
-import initSearchAutocompletes from "./components/search-autocomplete";
+import initSearchAutocompletes from "./components/form/search-autocomplete";
 import initToasts from "./components/toast";
 import FileDropzoneUpload from "./features/file-upload/file-dropzone-upload";
 import {initImageDropZones} from "./features/file-upload/image-dropzone-upload";
