+++ public/js/components/form/password-toggle.js
